pyproject.toml
[project]
name = "R2-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Guillermo A. Moran-Arreola",email = "g.a.moran.arreola@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django (>=6.0.3,<7.0.0)",
    "psycopg[binary] (>=3.3.3,<4.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "markdown (>=3.10.2,<4.0.0)",
    "django-cors-headers (>=4.9.0,<5.0.0)",
    "django-environ (>=0.11.0,<0.12.0) ; python_version < \"4\""
]

[tool.poetry]
packages = [{include = "core", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[dependency-groups]
dev = [
    "black (>=26.3.1,<27.0.0)",
    "isort (>=8.0.1,<9.0.0)",
    "flake8 (>=7.3.0,<8.0.0)"
]
